feat: implement chat functionality with socket integration and refactor components
