+++ src/app/(user)/chat/[id]/page.tsx
@@ -3,7 +3,6 @@ import ChatPage from "@pages/Member/Chat";
 import { ROUTES } from "@routes";
 import { getServerSession } from "next-auth";
 import { redirect } from "next/navigation";
-import Router from "next/router";
 
 export default async function Chat() {
